collapse multiple ::update_mailbox events. set proper date on MDN
* src/balsa-mblist.c: collapse multiple ::update_mailbox events.
* src/balsa-message.c: set proper date on MDN (Albrecht Dreß).
* src/mailbox-node.c: make sure "changed" is set for scanned mailboxes.
* libbalsa/{misc,address-book-vcard}.c: 64bit-fixes from FC2.
* libinit_balsa/balsa-druid-page-directory.c: spoolbox from FC2.

+++ libbalsa/address-book-vcard.c
@@ -285,7 +285,7 @@ load_vcard_file(LibBalsaAddressBook *ab)
 		address->address_list = g_list_reverse(address_list);
 
 		if (name) {
-                    gint bread, bwritten;
+                    gsize bread, bwritten;
                     GError *err = NULL;
                     gchar *cved = g_locale_to_utf8(name, -1, 
                                                    &bread, &bwritten, &err);
